refactor(Pawn): white와 black에 대한 상수 클래스를 생성 후 리팩토링
